- Linux running again.
git-svn-id: https://fuji.googlecode.com/svn/trunk@728 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/Util.h
@@ -40,7 +40,7 @@ inline void MFFixUp(T* &pPointer, void *pBase, int fix)
 {
 	if(pPointer)
 	{
-		int offset = (int&)pBase;
+		intp offset = (intp&)pBase;
 
 		if(!fix)
 			offset = -offset;
